% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modBioPed.R
\name{modBioPed}
\alias{modBioPed}
\title{The actual OpenMx model}
\usage{
modBioPed(nv, vrs, AModel, DModel, EmModel, EfModel, EnModel, EpModel, EcModel,
  MModel, data_M, data_F, data_PP, data_P, data_C)
}
\arguments{
\item{nv}{number of variables to be analysed}

\item{vrs}{string vector with names of all \eqn{p} outcome variables.}

\item{AModel}{list with OpenMx components for additive genetic model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VAf.}

\item{DModel}{list with OpenMx components for dominance genetic model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VD.}

\item{EmModel}{list of OpenMx components for mother environmental model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VEm.}

\item{EfModel}{list of OpenMx components for father environmental model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VEf.}

\item{EnModel}{list of OpenMx components for nuclear family environmental model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VEn.}

\item{EpModel}{list of OpenMx components for twin/pregnancy environmental model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VEp.}

\item{EcModel}{list of OpenMx components for unique environmental model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VEc.}

\item{MModel}{list of OpenMx components for mean model.
Must contain mxMatrix/mxAlgebra with dim \eqn{px1} named My.}

\item{data_M}{dataframe with mother identifiers.}

\item{data_F}{dataframe with father identifiers.}

\item{data_PP}{dataframe with parent-pair identifiers.}

\item{data_P}{dataframe with pregnancy identifiers and mz dummy.}

\item{data_C}{dataframe with pregnancy identifiers, mz dummy, covariates and variables to be modelled (vrs).}
}
\value{
fitted OpenMx model
}
\description{
The actual OpenMx model
}
