% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBioPed.R
\name{runBioPed}
\alias{runBioPed}
\title{runBioPed}
\usage{
runBioPed(vrs, AModel, DModel, EmModel, EfModel, EnModel, EpModel, EcModel,
  MModel, data_M, data_F, data_PP, data_P, data_C, ...)
}
\arguments{
\item{vrs}{string vector with names of all \eqn{p} outcome variables.}

\item{AModel}{list with OpenMx components for additive genetic model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VA.}

\item{DModel}{list with OpenMx components for dominance genetic model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VD.}

\item{EmModel}{list of OpenMx components for mother environmental model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VEm.}

\item{EfModel}{list of OpenMx components for father environmental model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VEf.}

\item{EnModel}{list of OpenMx components for nuclear family environmental model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VEn.}

\item{EpModel}{list of OpenMx components for twin/pregnancy environmental model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VEp.}

\item{EcModel}{list of OpenMx components for unique environmental model.
Must contain mxMatrix/mxAlgebra with dim \eqn{pxp} named VEc.}

\item{MModel}{list of OpenMx components for mean model.
Must contain mxMatrix/mxAlgebra with dim \eqn{px1} named My.}

\item{data_M}{dataframe with mother identifiers.}

\item{data_F}{dataframe with father identifiers.}

\item{data_PP}{dataframe with parent-pair identifiers.}

\item{data_P}{dataframe with pregnancy identifiers and mz dummy.}

\item{data_C}{dataframe with pregnancy identifiers, mz dummy, covariates and variables to be modelled (vrs).}

\item{...}{Arguments to be passed to \code{MxTryhard()}.}
}
\description{
Fits a biometric model based on parent identifiers.
}
\details{
The arguments \code{vrs, data_M, data_F, data_PP, data_P} and \code{data_C} must all be supplied.
In addition, at least one of the arguments \code{AModel, DModel, EpModel, EnModel, EcModel} or \code{MModel} must be supplied.
}
\examples{
library(OpenMx)
library(dplyr)
library(labOpenMx)

data(D_5var) # Load example dataframe

nv = 5
vrs = paste0("Y", 1:nv)

D_M = data.frame(Mid = unique(D_5var$Mid))
D_F = data.frame(Fid = unique(D_5var$Fid))
D_PP = data.frame(PPid = unique(D_5var$PPid))
D_P = D_5var \%>\%
 group_by(Pid) \%>\%
 summarise(PPid = unique(PPid),
           Mid = unique(Mid),
           Fid = unique(Fid),
           mz = unique(mz)) \%>\%
 ungroup() \%>\%
 data.frame()

D_C = D_5var \%>\%
 dplyr::select(Y1:Y5, female, mz, Pid) \%>\%
 data.frame()

# Fit an "independent pathway" model
# Additive genetic model
Amod = list(mxMatrix("Full", nv, 1, T, 0.5, labF("lCA", nv, 1), name = "lCA"),
           mxMatrix("Diag", nv, nv, T, 0.5, labD("lSA", nv), name = "lSA"),
           mxAlgebra(lCA\%*\%t(lCA) + lSA\%*\%t(lSA), name = "VA"))

# pregnancy/twin specific environmental model
Tmod = list(mxMatrix("Full", nv, 1, T, 0.5, labF("lCT", nv, 1), name = "lCT"),
           mxMatrix("Diag", nv, nv, T, 0.5, labD("lST", nv), name = "lST"),
           mxAlgebra(lCT\%*\%t(lCT) + lST\%*\%t(lST), name = "VEp"))

# Unique environmental model
Emod = list(mxMatrix("Full", nv, 1, T, 0.5, labF("lCE", nv, 1), name = "lCE"),
           mxMatrix("Diag", nv, nv, T, 0.5, labD("lSE", nv), name = "lSE"),
           mxAlgebra(lCE\%*\%t(lCE) + lSE\%*\%t(lSE), name = "VEc"))

# Mean model
Mmod = list(mxMatrix("Full", nv, 2, T, 0, labF("b", nv, 2), name = "B"),
           mxMatrix("Full", 2, 1, F, 1, c("x1", "data.female"), name = "X"),
           mxAlgebra(B\%*\%X, name = "My"))

fit = runBioPed(vrs, AModel = Amod, EpModel = Tmod, EcModel = Emod, MModel = Mmod,
               data_M = D_M, data_F = D_F, data_PP = D_PP, data_P = D_P, data_C = D_C)
summary(fit)

}
